Default llama_dtype='bf16' + plumb dtype across fp32/bf16 boundary
User confirmed production runs use bf16 Llama; switched the default so
callers don't need to opt in. The mixed-dtype stack (fp32 mem_inject +
memory params, bf16 Llama + memory state) now works without autocast.

Ownership split:

- **Llama weights**: bf16 (config default) or fp32 (bit-exact test opt-in).
- **MemInjectLayer W_in/W_out/scale**: fp32 nn.Parameter. bf16 params
  round small Adam updates to zero (smoke caught `scale` stuck at 2.0
  for a full optimizer step), and these three are tiny, so the fp32
  cost is negligible. Forward casts inputs to fp32 for the matmul,
  then casts the scaled residual back to hidden_states' dtype.
- **Memory graph params**: fp32 for the same optimizer reason.
- **Memory state** (h, msg, W, hebbian, prev_readout, s_mem_*):
  follows Llama's dtype via `reset_memory`. Matches the carried-state
  dtype to what h_mem enters as.
- **Memory buffers** (out_port_mask, role_id): unchanged dtypes.
  `fused_memory_step_torch` now casts out_port_mask to h_out dtype
  per call so `h_out * mask` stays in state dtype — was silently
  upcasting readout to fp32.
- **Modulator internal path**: stays in weight dtype (fp32) through
  build_tokens + AttnBlocks + pool_norm + logit_head; the state-dtype
  cast lands at modulator OUTPUT (emb lookup + decoder's
  `.to(W.dtype)` calls), not at each tokens.to() boundary. Removed
  the tokens / edge_bias cast-back that was producing bf16 inputs to
  fp32 SDPA.
- **LlamaMemAdapter.mem_head_logits**: casts readouts into W_out's
  dtype for the projection, then into Llama's dtype for norm + lm_head.

Also:
- `Config.initialize_states(dtype=None)`: accepts explicit dtype so
  the wrapper can force memory state to match Llama.
- `memory.forward_segment` only casts LM weights when dtype actually
  differs — was unconditionally allocating 200 MB bf16 copy of
  lm_head per call.

AR rollout bench on RTX 4090 (Llama-3B, BS=8, T_pre=128, T_gen=128):
    Vanilla bf16:                10.23 ms/gen-token
    Wrapper + memory:            11.30 ms/gen-token
    Memory overhead:              1.07 ms/gen-token (10%)

(Original pre-fix wrapper was 35.79 ms/gen-token = 251% overhead —
the dtype-plumbing + aux-loss gating changes deliver a 19.5 ms/token
speedup.)

32/32 CPU smoke tests still passing.

+++ src/pretrained/config.py
@@ -22,12 +22,13 @@ class PretrainedConfig:
     inject_layer: int = 8
     # Backbone freeze. The smoke test flips this off briefly; training stays on.
     freeze_backbone: bool = True
-    # Llama load dtype. "fp32" preserves bit-exact vanilla-Llama parity for
-    # unit tests. "bf16" matches CUDA autocast behavior and skips the
-    # on-the-fly fp32→bf16 weight cast that autocast triggers on every
-    # matmul (profile showed this was 75% of CUDA time at T=1 AR rollout).
-    # Callers on GPU production runs should pick "bf16".
-    llama_dtype: str = "fp32"
+    # Llama load dtype. "bf16" (default) matches production runs and skips
+    # the on-the-fly fp32→bf16 weight casts that autocast otherwise triggers
+    # on every matmul (profile showed these were 75% of CUDA time at T=1 AR
+    # rollout). The fp32 path is only needed for the bit-exact vanilla-Llama
+    # parity smoke test, which opts in explicitly via
+    # `PretrainedConfig.llama_1b(llama_dtype='fp32')`.
+    llama_dtype: str = "bf16"
     # LM hidden dim — populated from HF config at load time.
     d_lm: int = -1
     # LM num layers — populated from HF config at load time.
